+++ test/draw/Image.spec.ts
@@ -47,10 +47,6 @@ describe(Image.name, () => {
       image = document.addParagraph().addImage(join(__dirname, "..", "data", "ODF.png"));
     });
 
-    it("add draw namespace", () => {
-      expect(document.toString()).toMatch(/xmlns:draw="urn:oasis:names:tc:opendocument:xmlns:drawing:1.0"/);
-    });
-
     it("append a draw frame with image and base64 encoded image", () => {
       const regex = new RegExp("<draw:frame text:anchor-type=\"paragraph\">"
         + "<draw:image>"